Add files via upload
added in the program that can produce sound output. right now it pulls from a file called pmodSamples_full.txt and uses that data to change the output frequency of the sound. I am gong to split up pmodSamples_full into two different text files and use the data from one of them to control frequency, and the other one to control amplitude.